Fix the blocks count states and default values conflict for inherit from theme
