+++ src/blocks.js
@@ -167,31 +167,9 @@ import "./extensions/attributes.js";
 import "./utils/accordionOneOpen.js"
 import "./utils/responsiveTabSync.js"
 
-// Auto Block Recovery System
-import autoBlockRecovery from './utils/autoBlockRecovery';
-import initGlobalInheritFromTheme from './utils/globalInheritFromTheme';
+// Auto Block Recovery System and Inherit from Theme.
+import {initAutoBlockRecovery, initGlobalInheritFromThemeSystem} from './utils/settings.js'
 import domReady from '@wordpress/dom-ready';
-
-// Initialize Auto Block Recovery - Only if enabled in settings
-export const initAutoBlockRecovery = () => {
-	// Check if auto block recovery is enabled
-	const isAutoRecoveryEnabled = responsive_globals && responsive_globals.auto_block_recovery === '1';
-	
-	if ( isAutoRecoveryEnabled && window._wpLoadBlockEditor ) {
-		window._wpLoadBlockEditor.then( () => {
-			autoBlockRecovery();
-		} );
-	}
-};
-
-// Initialize Global Inherit From Theme functionality
-export const initGlobalInheritFromThemeSystem = () => {
-	if (window._wpLoadBlockEditor) {
-		window._wpLoadBlockEditor.then(() => {
-			initGlobalInheritFromTheme();
-		});
-	}
-};
 
 // Initialize auto block recovery only if enabled
 domReady( initAutoBlockRecovery );
