+++ src/blocks/popup/attributes.js
@@ -1,3 +1,5 @@
+import { convertTruthyFalsyValue } from "../../utils/helper";
+
 const attributes = {
     block_id: {
         type: "string",
@@ -794,7 +796,7 @@ const attributes = {
     },
     inheritFromTheme: {
       type: "boolean",
-      default: (typeof responsive_globals !== 'undefined' && responsive_globals.global_inherit_from_theme === '1'),
+      default: convertTruthyFalsyValue(responsive_globals?.global_inherit_from_theme),
     },
     inheritFromThemesaved: {
       type: "boolean",
